reconcile the two openapi folders
