+++ packages/openapi/package.json
@@ -7,6 +7,11 @@
       "bun": "./index.ts",
       "types": "./dist/index.d.ts",
       "import": "./dist/index.js"
+    },
+    "./browser": {
+      "bun": "./browser.ts",
+      "types": "./dist/browser.d.ts",
+      "import": "./dist/browser.js"
     }
   },
   "scripts": {
