+++ packages/source-stripe/src/browser.ts
@@ -1,0 +1,1 @@
+export * from '@stripe/sync-openapi/browser'
