+++ packages/source-stripe/src/index.ts
@@ -440,8 +440,8 @@ export default createStripeSource()
 export { subdivideRanges } from '@stripe/sync-protocol'
 export { buildResourceRegistry, DEFAULT_SYNC_OBJECTS, EXCLUDED_TABLES } from './resourceRegistry.js'
 export { catalogFromOpenApi } from './catalog.js'
-export { SpecParser, OPENAPI_RESOURCE_TABLE_ALIASES } from './openapi/specParser.js'
-export type { ParsedResourceTable, ParsedOpenApiSpec } from './openapi/types.js'
+export { SpecParser, OPENAPI_RESOURCE_TABLE_ALIASES } from '@stripe/sync-openapi'
+export type { ParsedResourceTable, ParsedOpenApiSpec } from '@stripe/sync-openapi'
 export type { RateLimiter } from './rate-limiter.js'
 export { createInMemoryRateLimiter } from './rate-limiter.js'
 export { verifyWebhookSignature, WebhookSignatureError } from './webhookVerify.js'
