--- packages/source-stripe/src/openapi/browser.ts
@@ -1,2 +1,0 @@
-export { SpecParser, OPENAPI_RESOURCE_TABLE_ALIASES } from './specParser.js'
-export type { ParsedResourceTable, ParsedOpenApiSpec } from './types.js'
