--- packages/source-stripe/src/openapi/index.ts
@@ -1,9 +1,0 @@
-export type * from './types.js'
-export {
-  SpecParser,
-  RUNTIME_REQUIRED_TABLES,
-  OPENAPI_RESOURCE_TABLE_ALIASES,
-  RUNTIME_RESOURCE_ALIASES,
-} from './specParser.js'
-export { OPENAPI_COMPATIBILITY_COLUMNS } from './runtimeMappings.js'
-export { parsedTableToJsonSchema } from './jsonSchemaConverter.js'
